+++ src/dataspec/base.py
@@ -137,7 +137,7 @@ class Spec(ABC):
     """
     The abstract base class of all Specs.
 
-    Any Spec returned from :py:data:`dataspec.s` will conform to this interface.
+    All Specs returned by :py:data:`dataspec.s` conform to this interface.
     """
 
     @property
@@ -150,39 +150,43 @@ class Spec(ABC):
         """
         raise NotImplementedError
 
-    @property
-    @abstractmethod
-    def _conformer(self) -> Optional[Conformer]:  # pragma: no cover
-        """Return the custom conformer for this Spec."""
-        raise NotImplementedError
-
-    @property
-    def conformer(self) -> Conformer:
-        """Return the conformer attached to this Spec."""
-        return self._conformer or self._default_conform
-
     @abstractmethod
     def validate(self, v: Any) -> Iterator[ErrorDetails]:  # pragma: no cover
         """
-        Validate the value ``v`` against the Spec.
+        Validate the value ``v`` against the Spec, yielding successive Spec failures as
+        :py:class:`dataspec.ErrorDetails` instances, if any.
 
-        Yields an iterable of Spec failures, if any.
+        By definition, if ``next(spec.validate(v))`` raises ``StopIteration``, the
+        first time it is called, the value is considered valid according to the Spec.
+
+        :param v: a value to validate
+        :return: an iterator of Spec failures as :py:class:`dataspec.ErrorDetails`
+            instances, if any
         """
         raise NotImplementedError
 
-    def validate_ex(self, v: Any):
+    def validate_ex(self, v: Any) -> None:
         """
-        Validate the value ``v`` against the Spec.
+        Validate the value ``v`` against the Spec, throwing a
+        :py:class:`dataspec.ValidationError` containing a list of all of the Spec
+        failures for ``v`` , if any. Returns :py:obj:`None` otherwise.
 
-        Throws a ``ValidationError`` with a list of Spec failures, if any.
+        :param v: a value to validate
+        :return: :py:obj:`None`
         """
         errors = list(self.validate(v))
         if errors:
             raise ValidationError(errors)
 
     def is_valid(self, v: Any) -> bool:
-        """Returns True if ``v`` is valid according to the Spec, otherwise returns
-        False."""
+        """
+        Returns :py:obj:`True` if ``v`` is valid according to the Spec, otherwise
+        returns :py:obj:`False`.
+
+        :param v: a value to validate
+        :return: :py:obj:`True` if the value is valid according to the Spec, otherwise
+            :py:obj:`False`
+        """
         try:
             next(self.validate(v))
         except StopIteration:
@@ -190,37 +194,103 @@ class Spec(ABC):
         else:
             return False
 
-    def _default_conform(self, v):
-        """
-        Default conformer for the Spec.
+    @property
+    def conformer(self) -> Optional[Conformer]:  # pragma: no cover
+        """Return the custom conformer attached to this Spec, if one is defined."""
+        return None
 
-        If no custom conformer is specified, this conformer will be invoked.
+    def conform(self, v: Any):
         """
-        return v
+        Conform ``v`` to the Spec, returning the possibly conformed value or an
+        instance of :py:class:`dataspec.Invalid` if the value is invalid cannot
+        be conformed.
 
-    def conform_valid(self, v):
+        Exceptions arising from calling :py:attr:`dataspec.Spec.conformer` with ``v``
+        will be raised from this method.
+
+        :param v: a value to conform
+        :return: a conformed value or a :py:class:`dataspec.Invalid` instance if the
+            input value could not be conformed
+        """
+        if self.is_valid(v):
+            return self.conform_valid(v)  # pylint: disable=not-callable
+        else:
+            return INVALID
+
+    def conform_valid(self, v: Any):
         """
         Conform ``v`` to the Spec without checking if v is valid first and return the
         possibly conformed value or ``INVALID`` if the value cannot be conformed.
 
         This function should be used only if ``v`` has already been check for validity.
+
+        Exceptions arising from calling :py:attr:`dataspec.Spec.conformer` with ``v``
+        will be raised from this method.
+
+        :param v: a *validated* value to conform
+        :return: a conformed value or a :py:class:`dataspec.Invalid` instance if the
+            input value could not be conformed
         """
-        return self.conformer(v)
+        if self.conformer is None:
+            return v
+        return self.conformer(v)  # pylint: disable=not-callable
 
-    def conform(self, v: Any):
-        """Conform ``v`` to the Spec, first checking if v is valid, and return the
-        possibly conformed value or ``INVALID`` if the value cannot be conformed."""
-        if self.is_valid(v):
-            return self.conform_valid(v)
-        else:
-            return INVALID
+    def compose_conformer(self, conformer: Conformer) -> "Spec":
+        """
+        Return a new Spec instance with a new conformer which is the composition of the
+        ``conformer`` and the current conformer for this Spec instance.
 
-    def with_conformer(self, conformer: Conformer) -> "Spec":
-        """Return a new Spec instance with the new conformer."""
+        If the current Spec instance has a custom conformer, this is equivalent to
+        calling ``spec.with_conformer(lambda v: conformer(spec.conformer(v)))``.
+        If the current Spec instance has no custom conformer, this is equivalent to
+        calling :py:meth:`dataspec.Spec.with_conformer` with ``conformer``.
+
+        To completely replace the conformer for this Spec instance, use
+        :py:meth:`dataspec.Spec.with_conformer`.
+
+        This method does not modify the current Spec instance.
+
+        :param conformer: a conformer to compose with the conformer of the current
+            Spec instance
+        :return: a copy of the current Spec instance with the new composed conformer
+        """
+        existing_conformer = self.conformer
+
+        if existing_conformer is None:
+            return self.with_conformer(conformer)
+
+        def conform_spec(v: T) -> Union[V, Invalid]:
+            assert existing_conformer is not None
+            return conformer(existing_conformer(v))  # pylint: disable=not-callable
+
+        return self.with_conformer(conform_spec)
+
+    def with_conformer(self, conformer: Optional[Conformer]) -> "Spec":
+        """
+        Return a new Spec instance with the new conformer, replacing any custom
+        conformers.
+
+        If ``conformer`` is :py:obj:`None` , the returned Spec will have no custom
+        conformer.
+
+        To return a copy of the current Spec with a composition of the current
+        Spec instance, use :py:meth:`dataspec.Spec.compose_conformer`.
+
+        :param conformer: a conformer to replace the conformer of the current Spec
+            instance or :py:obj:`None` to remove the conformer associated with this
+        :return: a copy of the current Spec instance with new conformer
+        """
         return attr.evolve(self, conformer=conformer)
 
     def with_tag(self, tag: Tag) -> "Spec":
-        """Return a new Spec instance with the new tag applied."""
+        """
+        Return a new Spec instance with the new tag applied.
+
+        This method does not modify the current Spec instance.
+
+        :param tag: a new tag to use for the new Spec
+        :return: a copy of the current Spec instance with the new tag applied
+        """
         return attr.evolve(self, tag=tag)
 
 
@@ -232,7 +302,7 @@ class ValidatorSpec(Spec):
 
     tag: Tag
     _validate: ValidatorFn
-    _conformer: Optional[Conformer] = None
+    conformer: Optional[Conformer] = None
 
     def validate(self, v) -> Iterator[ErrorDetails]:
         try:
@@ -291,7 +361,7 @@ class PredicateSpec(Spec):
 
     tag: Tag
     _pred: PredicateFn
-    _conformer: Optional[Conformer] = None
+    conformer: Optional[Conformer] = None
 
     def validate(self, v) -> Iterator[ErrorDetails]:
         try:
@@ -312,7 +382,7 @@ class PredicateSpec(Spec):
 class CollSpec(Spec):
     tag: Tag
     _spec: Spec
-    _conformer: Optional[Conformer] = None
+    conformer: Optional[Conformer] = None
     _out_type: Optional[Type] = None
     _validate_coll: Optional[ValidatorSpec] = None
 
@@ -427,10 +497,13 @@ class CollSpec(Spec):
         if validators:
             validate_coll = ValidatorSpec.from_validators("coll", *validators)
 
+        def conform_coll(v: Iterable) -> Iterable:
+            return (out_type or type(v))(spec.conform(e) for e in v)  # type: ignore[call-arg]  # noqa
+
         return cls(
             tag or "coll",
             spec=spec,
-            conformer=conformer,
+            conformer=compose_conformers(conform_coll, *filter(None, (conformer,))),
             out_type=out_type,
             validate_coll=validate_coll,
         )
@@ -441,9 +514,6 @@ class CollSpec(Spec):
 
         for i, e in enumerate(v):
             yield from _enrich_errors(self._spec.validate(e), self.tag, i)
-
-    def _default_conform(self, v):
-        return (self._out_type or type(v))(self._spec.conform(e) for e in v)
 
 
 # In Python 3.6, you cannot inherit directly from Generic with slotted classes:
@@ -458,14 +528,14 @@ class DictSpec(Spec):
     tag: Tag
     _reqkeyspecs: Mapping[Any, Spec] = attr.ib(factory=dict)
     _optkeyspecs: Mapping[Any, Spec] = attr.ib(factory=dict)
-    _conformer: Optional[Conformer] = None
+    conformer: Optional[Conformer] = None
 
     @classmethod
     def from_val(
         cls,
         tag: Optional[Tag],
         kvspec: Mapping[str, SpecPredicate],
-        conformer: Conformer = None,
+        conformer: Optional[Conformer] = None,
     ):
         reqkeys = {}
         optkeys: MutableMapping[Any, Spec] = {}
@@ -475,8 +545,22 @@ class DictSpec(Spec):
             else:
                 reqkeys[k] = make_spec(v)
 
+        def conform_mapping(d: Mapping) -> Mapping:
+            conformed_d = {}
+            for k, spec in reqkeys.items():
+                conformed_d[k] = spec.conform(d[k])
+
+            for k, spec in optkeys.items():
+                if k in d:
+                    conformed_d[k] = spec.conform(d[k])
+
+            return conformed_d
+
         return cls(
-            tag or "map", reqkeyspecs=reqkeys, optkeyspecs=optkeys, conformer=conformer
+            tag or "map",
+            reqkeyspecs=reqkeys,
+            optkeyspecs=optkeys,
+            conformer=compose_conformers(conform_mapping, *filter(None, (conformer,))),
         )
 
     def validate(self, d) -> Iterator[ErrorDetails]:  # pylint: disable=arguments-differ
@@ -505,19 +589,23 @@ class DictSpec(Spec):
             if k in d:
                 yield from _enrich_errors(vspec.validate(d[k]), self.tag, k)
 
-    def _default_conform(self, d):  # pylint: disable=arguments-differ
-        conformed_d = {}
-        for k, spec in self._reqkeyspecs.items():
-            conformed_d[k] = spec.conform(d[k])
-
-        for k, spec in self._optkeyspecs.items():
-            if k in d:
-                conformed_d[k] = spec.conform(d[k])
-
-        return conformed_d
-
 
 class ObjectSpec(DictSpec):
+    @classmethod
+    def from_val(
+        cls,
+        tag: Optional[Tag],
+        kvspec: Mapping[str, SpecPredicate],
+        conformer: Optional[Conformer] = None,
+    ):
+        """
+        Return a Spec for an arbitrary object instance.
+
+        Overwrite the default conformer provided for ``DictSpec`` s, since it does not
+        make sense for objects.
+        """
+        return super().from_val(tag, kvspec).with_conformer(conformer)
+
     def validate(self, o) -> Iterator[ErrorDetails]:  # pylint: disable=arguments-differ
         for k, vspec in self._reqkeyspecs.items():
             if hasattr(o, k):
@@ -534,9 +622,6 @@ class ObjectSpec(DictSpec):
         for k, vspec in self._optkeyspecs.items():
             if hasattr(o, k):
                 yield from _enrich_errors(vspec.validate(getattr(o, k)), self.tag, k)
-
-    def _default_conform(self, o):  # pylint: disable=arguments-differ
-        raise TypeError("Cannot use a default conformer for an Object")
 
 
 def _enum_conformer(e: EnumMeta) -> Conformer:
@@ -559,7 +644,7 @@ def _enum_conformer(e: EnumMeta) -> Conformer:
 class SetSpec(Spec):
     tag: Tag
     _values: Union[Set, FrozenSet]
-    _conformer: Optional[Conformer] = None
+    conformer: Optional[Conformer] = None
 
     def validate(self, v) -> Iterator[ErrorDetails]:
         if v not in self._values:
@@ -570,13 +655,29 @@ class SetSpec(Spec):
                 via=[self.tag],
             )
 
+    @classmethod
+    def from_enum(
+        cls, tag: Optional[Tag], pred: EnumMeta, conformer: Optional[Conformer] = None
+    ):
+        return cls(
+            tag or pred.__name__,
+            frozenset(
+                chain.from_iterable(
+                    [mem, mem.name, mem.value] for mem in pred  # type: ignore[var-annotated]  # noqa
+                )
+            ),
+            conformer=compose_conformers(
+                _enum_conformer(pred), *filter(None, (conformer,)),
+            ),
+        )
+
 
 @attr.s(auto_attribs=True, frozen=True, slots=True)
 class TupleSpec(Spec):
     tag: Tag
     _pred: Tuple[SpecPredicate, ...]
     _specs: Tuple[Spec, ...]
-    _conformer: Optional[Conformer] = None
+    conformer: Optional[Conformer] = None
     _namedtuple: Optional[Type[NamedTuple]] = None
 
     @classmethod
@@ -596,11 +697,16 @@ class TupleSpec(Spec):
         else:
             namedtuple_type = None  # type: ignore
 
+        def conform_tuple(v) -> Union[Tuple, NamedTuple]:
+            return ((namedtuple_type and namedtuple_type._make) or tuple)(
+                spec.conform(v) for spec, v in zip(specs, v)
+            )
+
         return cls(
             tag or "tuple",
             pred=pred,
             specs=specs,
-            conformer=conformer,
+            conformer=compose_conformers(conform_tuple, *filter(None, (conformer,))),
             namedtuple=namedtuple_type,  # type: ignore
         )
 
@@ -621,11 +727,6 @@ class TupleSpec(Spec):
             yield ErrorDetails(
                 message=f"Value is not a tuple type", pred=self, value=t, via=[self.tag]
             )
-
-    def _default_conform(self, v):
-        return ((self._namedtuple and self._namedtuple._make) or tuple)(
-            spec.conform(v) for spec, v in zip(self._specs, v)
-        )
 
 
 def _enrich_errors(
@@ -657,7 +758,7 @@ def compose_conformers(*conformers: Conformer) -> Conformer:
         conformed_v = v
         for conform in conformers:
             conformed_v = conform(conformed_v)
-            if conformed_v is INVALID:
+            if conformed_v is INVALID or isinstance(conformed_v, Invalid):
                 break
         return conformed_v
 
@@ -673,17 +774,14 @@ def compose_spec_conformers(
 
     Apply the ``conform_final`` conformer on the final return from the composition, if
     any.
+
+    This function bypasses the :py:meth:`dataspec.Spec.conform` method and accesses
+    :py:attr:`dataspec.Spec.conformer` directly.
     """
 
-    def do_conform(v):
-        conformed_v = v
-        for spec in specs:
-            conformed_v = spec.conform(conformed_v)
-            if conformed_v is INVALID:
-                break
-        return conformed_v if conform_final is None else conform_final(conformed_v)
-
-    return do_conform
+    return compose_conformers(
+        *filter(None, chain((spec.conformer for spec in specs), (conform_final,)))
+    )
 
 
 _MUNGE_NAMES = re.compile(r"[\s|-]")
@@ -785,7 +883,7 @@ def make_spec(  # pylint: disable=inconsistent-return-statements  # noqa: MC0001
 
     Specs may be created from Python types, in which case a Spec will be produced
     that performs an :py:func:`isinstance` check. :py:obj:`None` may be provided as
-    a shortcut for ``type(None)`` . To specify a nilable value, you should use
+    a shortcut for ``type(None)``. To specify a nilable value, you should use
     :py:meth:`dataspec.SpecAPI.nilable` instead.
 
     Specs may be created for enumerated types using a Python ``set`` or ``frozenset``
@@ -816,6 +914,14 @@ def make_spec(  # pylint: disable=inconsistent-return-statements  # noqa: MC0001
     the input tuple being validated and conformed to the corresponding element in the
     Spec.
 
+    Specs may be be created from existing Specs. If an existing :py:class:`datspec.Spec`
+    instance is given, that Spec will be returned without modification. If a tag is
+    given, a new Spec will be created from the existing Spec with the new tag. If a
+    conformer is given, a new Spec will be created from the existing Spec with the new
+    conformer (*replacing* any conformer on the existing Spec, rather than composing).
+    If both a new tag and conformer are given, a new Spec will be returned with both
+    the new tag and conformer.
+
     :param tag: an optional :py:data:`dataspec.Tag` for the resulting spec
     :param pred: a value which can be be converted into a :py:class:`dataspec.Spec`
     :param conformer: an optional :py:data:`dataspec.Conformer` for the value
@@ -831,11 +937,7 @@ def make_spec(  # pylint: disable=inconsistent-return-statements  # noqa: MC0001
     if isinstance(pred, (frozenset, set)):
         return SetSpec(tag or "set", pred, conformer=conformer)
     elif isinstance(pred, EnumMeta):
-        return SetSpec(
-            tag or pred.__name__,
-            frozenset(chain.from_iterable([mem, mem.name, mem.value] for mem in pred)),
-            conformer=conformer or _enum_conformer(pred),
-        )
+        return SetSpec.from_enum(tag, pred, conformer=conformer)
     elif isinstance(pred, tuple):
         return TupleSpec.from_val(tag, pred, conformer=conformer)
     elif isinstance(pred, list):